+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # Changelog
 
-# 0.1.1
+## 0.2.0
 
-# 0.1.0
+### Feature
 
+- The preview of the original note is now displayed in the note create dialog
+- The username mention is now automatically added to the reply dialog
+
+### Fixed
+
+- Some mention avatars on the notes page don't display in some cases
+- The link for MFM is not opening
+
+### Other
+
+- Upgrade flutter dependencies